test: view 단위 테스트 추가

+++ src/main/java/lotto/dto/WinningStatisticsDto.java
@@ -1,5 +1,5 @@
 package lotto.dto;
 
 public record WinningStatisticsDto(int threeMatches, int fourMatches, int fiveMatches,
-                                   int fiveMatchesWithBonus, int sixMatches, float earningRate) {
+                                   int fiveMatchesWithBonus, int sixMatches, double earningRate) {
 }
